Always black borders for focused TextField
If you put in one number (1 for example), the TextFields border would remain thicker, but turn grey - even tough it's still in focus. I don't think this was intended behaviour, and it's not something I can find in the design.

Since 'focusedBorder' is only used when the TextField is focused, we don't also need to look at the focusNode to determine color here.

I don't know why it caused an issue, I would have thought it would always have focus if the border 'focusedBorder' is being used...