Changed onPanDown to onTapDown
Because previously onPanUpdate was triggered on the medium tablets directly after onPanDown, even if the pointer didn't move, which meant that you could not select a 5 min interval by just tapping. This change means that onPanUpdate isn't called unless the pointer actually moved. Not sure why this works and the previous solution didn't. It worked on the emulators, but not on the medium tablet device.